requirements.txt
aiogram==2.25.2
mysql-connector-python==8.4.0
loguru==0.6.0


